+++ src/types/comment.ts
@@ -1,0 +1,5 @@
+export default interface comment {
+    body: string;
+    timeStamp: any;
+    userHandle: string;
+  }
