+++ src/types/login.ts
@@ -1,0 +1,4 @@
+export default interface userLoginData {
+    password: string;
+    email: string;
+  }
