+++ src/types/post.ts
@@ -1,0 +1,11 @@
+export default interface onePost {
+    postID: string;
+    userHandle: string;
+    content: string;
+    image: string;
+    timeStamp: any;
+    likes: number;
+    dislikes: number;
+    comments: number;
+    postInterests: Array<string>;
+  }
