+++ src/types/user.ts
@@ -1,0 +1,9 @@
+export default interface userData {
+    email: string;
+    password: string;
+    confirmPassword: string;
+    userHandle: string;
+    userImage: string;
+    userInterests: Array<string>;
+  }
+
